move data request to the real module
